MC-222 running nginx on port 80 working, logstash working
I found it problematic to remove apache entirely, because the eb
deploy and status checks seem to rely on it.  However, I minimized
the resources it uses, configured logstash to pull logs from
nginx, and pulled apache completely out of the request processing.
So, this seems like a good intermediate step to make available.