MC-268 correcting hero image SQL
The previous SQL for the hero image was just wrong.  This is ugly,
but works correctly.

Getting the hero image via a correlated subquery isn't that bad, despite
the fact that correlated subqueries generally aren't recommended.  However,
since multiple columns of the hero image are needed, I need to repeat the
subquery four times.  There doesn't seem to be another query syntax that
MySQL will buy.

Restructuring the query to get the hero image information in a second API
call would allow for a better structuring of the query and perform better
(this doesn't seem to be terrible, surprisingly).